collector/requirements.txt
flask==3.1.1
pyyaml==6.0.2
paramiko==3.5.1
